Added newline at end of file when needed. Added casting in VoxelRGB/HSV operators.
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/soma/soma-io/trunk@57104 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/libcartobase/type/voxelrgb_def.h
@@ -222,9 +222,9 @@ namespace carto {
     #ifdef CARTO_DEBUG_RGB
       std::cout << "RGB:: operator *= ( uint8_t )" << std::endl;
     #endif
-    red()   *= value;
-    green() *= value;
-    blue()  *= value;
+    red()   *= (uint8_t) value;
+    green() *= (uint8_t) value;
+    blue()  *= (uint8_t) value;
     return *this;
   }
 
@@ -234,9 +234,9 @@ namespace carto {
     #ifdef CARTO_DEBUG_RGB
       std::cout << "RGB:: operator *= ( uint16_t )" << std::endl;
     #endif
-    red()   *= value;
-    green() *= value;
-    blue()  *= value;
+    red()   *= (uint8_t) value;
+    green() *= (uint8_t) value;
+    blue()  *= (uint8_t) value;
     return *this;
   }
 
@@ -246,9 +246,9 @@ namespace carto {
     #ifdef CARTO_DEBUG_RGB
       std::cout << "RGB:: operator *= ( uint32_t )" << std::endl;
     #endif
-    red()   *= value;
-    green() *= value;
-    blue()  *= value;
+    red()   *= (uint8_t) value;
+    green() *= (uint8_t) value;
+    blue()  *= (uint8_t) value;
     return *this;
   }
 
@@ -258,9 +258,9 @@ namespace carto {
     #ifdef CARTO_DEBUG_RGB
       std::cout << "RGB:: operator *= ( uint64_t )" << std::endl;
     #endif
-    red()   *= value;
-    green() *= value;
-    blue()  *= value;
+    red()   *= (uint8_t) value;
+    green() *= (uint8_t) value;
+    blue()  *= (uint8_t) value;
     return *this;
   }
 
@@ -270,9 +270,9 @@ namespace carto {
     #ifdef CARTO_DEBUG_RGB
       std::cout << "RGB:: operator *= ( float )" << std::endl;
     #endif
-    red()   *= value;
-    green() *= value;
-    blue()  *= value;
+    red()   *= (uint8_t) value;
+    green() *= (uint8_t) value;
+    blue()  *= (uint8_t) value;
     return *this;
   }
 
@@ -282,9 +282,9 @@ namespace carto {
     #ifdef CARTO_DEBUG_RGB
       std::cout << "RGB:: operator *= ( double )" << std::endl;
     #endif
-    red()   *= value;
-    green() *= value;
-    blue()  *= value;
+    red()   *= (uint8_t) value;
+    green() *= (uint8_t) value;
+    blue()  *= (uint8_t) value;
     return *this;
   }
 
@@ -295,9 +295,9 @@ namespace carto {
       std::cout << "RGB:: operator /= ( uint8_t )" << std::endl;
     #endif
     ASSERT( value != 0 );
-    red()   /= value;
-    green() /= value;
-    blue()  /= value;
+    red()   /= (uint8_t) value;
+    green() /= (uint8_t) value;
+    blue()  /= (uint8_t) value;
     return *this;
   }
 
@@ -308,9 +308,9 @@ namespace carto {
       std::cout << "RGB:: operator /= ( uint16_t )" << std::endl;
     #endif
     ASSERT( value != 0 );
-    red()   /= value;
-    green() /= value;
-    blue()  /= value;
+    red()   /= (uint8_t) value;
+    green() /= (uint8_t) value;
+    blue()  /= (uint8_t) value;
     return *this;
   }
 
@@ -321,9 +321,9 @@ namespace carto {
       std::cout << "RGB:: operator /= ( uint32_t )" << std::endl;
     #endif
     ASSERT( value != 0 );
-    red()   /= value;
-    green() /= value;
-    blue()  /= value;
+    red()   /= (uint8_t) value;
+    green() /= (uint8_t) value;
+    blue()  /= (uint8_t) value;
     return *this;
   }
 
@@ -334,9 +334,9 @@ namespace carto {
       std::cout << "RGB:: operator /= ( uint64_t )" << std::endl;
     #endif
     ASSERT( value != 0 );
-    red()   /= value;
-    green() /= value;
-    blue()  /= value;
+    red()   /= (uint8_t) value;
+    green() /= (uint8_t) value;
+    blue()  /= (uint8_t) value;
     return *this;
   }
 
@@ -347,9 +347,9 @@ namespace carto {
       std::cout << "RGB:: operator /= ( float )" << std::endl;
     #endif
     ASSERT( value != 0 );
-    red()   /= value;
-    green() /= value;
-    blue()  /= value;
+    red()   /= (uint8_t) value;
+    green() /= (uint8_t) value;
+    blue()  /= (uint8_t) value;
     return *this;
   }
 
@@ -360,9 +360,9 @@ namespace carto {
       std::cout << "RGB:: operator /= ( double )" << std::endl;
     #endif
     ASSERT( value != 0 );
-    red()   /= value;
-    green() /= value;
-    blue()  /= value;
+    red()   /= (uint8_t) value;
+    green() /= (uint8_t) value;
+    blue()  /= (uint8_t) value;
     return *this;
   }
 
@@ -596,9 +596,9 @@ namespace carto {
     #ifdef CARTO_DEBUG_RGB
       std::cout << "RGB:: operator *= ( long )" << std::endl;
     #endif
-    red()   *= value;
-    green() *= value;
-    blue()  *= value;
+    red()   *= (uint8_t) value;
+    green() *= (uint8_t) value;
+    blue()  *= (uint8_t) value;
     return *this;
   }
   
@@ -609,9 +609,9 @@ namespace carto {
       std::cout << "RGB:: operator /= ( long )" << std::endl;
     #endif
     ASSERT( value != 0 );
-    red()   /= value;
-    green() /= value;
-    blue()  /= value;
+    red()   /= (uint8_t) value;
+    green() /= (uint8_t) value;
+    blue()  /= (uint8_t) value;
     return *this;
   }
